Restructure folders, add blueprints, quick fixes
