+++ .gitignore
@@ -1,3 +1,5 @@
 venv
 .env
 __pycache__
+.pytest_cache
+setbox
