+++ app/schedule/routes.py
@@ -1,0 +1,7 @@
+from flask import Blueprint, render_template
+
+blueprint = Blueprint('schedule', __name__)
+
+@blueprint.route('/schedule')
+def schedule():
+    return render_template('schedule/schedule.html')
