+++ app/static/css/lesson.css
@@ -1,0 +1,10 @@
+.back-to-subject {
+    position: fixed;
+    top: 20px;
+    left: 20px;
+    background-color: #4CAF50;
+    color: white;
+    padding: 10px 20px;
+    border-radius: 5px;
+    cursor: pointer;
+}
