+++ app/templates/lesson/lesson.html
@@ -5,7 +5,7 @@
 {% endblock %}
 
 {% block main %}
-<a href="/" class="back-to-subject">Back to subject</a>
+<a href="/subject/{{ info[3] }}" class="back-to-subject">Back to subject</a>
 {% if info[1] == None %}
     <h2>{{ info[0] }} - {{ info[2]}}</h2>
 {% else %}
