--- environment.env
@@ -1,1 +1,0 @@
-FLASK_DEBUG = True
