+++ run.py
@@ -1,0 +1,4 @@
+from app.app import app
+
+if __name__ == '__main__':
+    app.run(host='localhost', port=3000, debug=True)
