+++ test_app.py
@@ -1,0 +1,9 @@
+import pytest
+
+
+@pytest.fixture
+def client():
+  app = app
+
+  with app.app_context():
+    yield app.test_client()
